+++ test/karma/fhir-sdc.spec.js
@@ -572,10 +572,28 @@ for (var i=0, len=fhirVersions.length; i<len; ++i) {
             });
           });
 
+          it('should preserve rendering-markdown extension on _text', function (){
+            $.get('test/data/R4/q-with-rendering-markdown-text.json', function(json) {
+              const lfData = fhir.SDC.convertQuestionnaireToLForms(json);
+              const qData = fhir.SDC.convertLFormsToQuestionnaire(lfData);
+              assert.ok(qData.item[0]._text);
+              assert.deepEqual(qData.item[0]._text, json.item[0]._text);
+            });
+          });
+
           it('should preserve rendering-xhtml extension on _prefix', function (){
             $.get('test/data/R4/q-with-rendering-xhtml-prefix.json', function(json) {
               const lfData = fhir.SDC.convertQuestionnaireToLForms(json);
               assert.equal(lfData.items[0]._prefixHTML, "<i class='testPlease'>A</i> HTML <b>prefix:</b>");
+              const qData = fhir.SDC.convertLFormsToQuestionnaire(lfData);
+              assert.ok(qData.item[0]._prefix);
+              assert.deepEqual(qData.item[0]._prefix, json.item[0]._prefix);
+            });
+          });
+
+          it('should preserve rendering-markdown extension on _prefix', function (){
+            $.get('test/data/R4/q-with-rendering-markdown-prefix.json', function(json) {
+              const lfData = fhir.SDC.convertQuestionnaireToLForms(json);
               const qData = fhir.SDC.convertLFormsToQuestionnaire(lfData);
               assert.ok(qData.item[0]._prefix);
               assert.deepEqual(qData.item[0]._prefix, json.item[0]._prefix);
@@ -660,6 +678,15 @@ for (var i=0, len=fhirVersions.length; i<len; ++i) {
               });
             });
 
+            it('should preserve rendering-markdown extension on answerOption _valueString', function (){
+              $.get('test/data/R4/q-with-rendering-markdown-answerOption.json', function(json) {
+                const lfData = fhir.SDC.convertQuestionnaireToLForms(json);
+                const qData = fhir.SDC.convertLFormsToQuestionnaire(lfData);
+                assert.ok(qData.item[0].item[0].answerOption[0]._valueString);
+                assert.deepEqual(qData.item[0].item[0].answerOption[0]._valueString, json.item[0].item[0].answerOption[0]._valueString);
+              });
+            });
+
             it('should preserve rendering-xhtml extension on answerOption valueCoding._display', function (){
               $.get('test/data/R4/q-with-rendering-xhtml-answerOption.json', function(json) {
                 const lfData = fhir.SDC.convertQuestionnaireToLForms(json);
@@ -668,6 +695,24 @@ for (var i=0, len=fhirVersions.length; i<len; ++i) {
                 const qData = fhir.SDC.convertLFormsToQuestionnaire(lfData);
                 assert.ok(qData.item[3].item[0].answerOption[0].valueCoding._display);
                 assert.deepEqual(qData.item[3].item[0].answerOption[0].valueCoding._display, json.item[3].item[0].answerOption[0].valueCoding._display);
+              });
+            });
+
+            it('should preserve rendering-markdown extension on answerOption valueCoding._display', function (){
+              $.get('test/data/R4/q-with-rendering-markdown-answerOption.json', function(json) {
+                const lfData = fhir.SDC.convertQuestionnaireToLForms(json);
+                const qData = fhir.SDC.convertLFormsToQuestionnaire(lfData);
+                assert.ok(qData.item[3].item[0].answerOption[0].valueCoding._display);
+                assert.deepEqual(qData.item[3].item[0].answerOption[0].valueCoding._display, json.item[3].item[0].answerOption[0].valueCoding._display);
+              });
+            });
+
+            it('should preserve rendering-markdown extension on help and legal', function (){
+              $.get('test/data/R4/q-with-rendering-markdown-help-and-legal.json', function(json) {
+                const lfData = fhir.SDC.convertQuestionnaireToLForms(json);
+                const qData = fhir.SDC.convertLFormsToQuestionnaire(lfData);
+                assert.deepEqual(qData.item[0].item[0]._text, json.item[0].item[0]._text);
+                assert.deepEqual(qData.item[1].item[0]._text, json.item[1].item[0]._text);
               });
             });