+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
+## [41.1.0] 2026-03-30
+### Added
+- Support for rendering-markdown extension. Note that rendering-xhtml
+  takes precedence over rendering-markdown if both are present.
+
 ## [41.0.0] 2026-03-27
 ### Changed
 - If answerExpression updates the answer list, we do not reset item.value